+++ app/src/main/java/com/br/imobiliaria/models/Filtro.java
@@ -2,12 +2,12 @@ package com.br.imobiliaria.models;
 
 public class Filtro {
     public int quantidadeQuartos;
-    public int preco;
+    public Double preco;
     public String localidade;
 
 
 
-    public Filtro(int quantidadeQuartos, int preco, String localidade) {
+    public Filtro(int quantidadeQuartos, Double preco, String localidade) {
         this.quantidadeQuartos = quantidadeQuartos;
         this.preco = preco;
         this.localidade = localidade;
